Move posterior caching to GPflowPredictor for happy hrvoje
